Create R_Cache_Implementation
Matrix Inverse is a Computationally Expensive Operation. 
Following Pair of functions , cache the inverse of a matrix and retrieve it from the cache if the same matrix inverse is retrieved by user. 
For a new unseen matrix, the inverse is newly calculated and pushed into the aforementioned cache

Similar functions can be implemented for other computationally expensive functions
This code is an enhancement to Prof. Roger Peng's (Johns Hopkins University) code and follows his framework